refactor: moved utils
